Reworked stories and added chromatic test for naked buttons
